dra-mobile-auth/dra_mobile_auth/requirements.txt
django
gunicorn
django-allauth
djangorestframework
djangorestframework-simplejwt
django-cors-headers
dj_rest_auth
phonenumbers
python-dotenv

